backend/trend-analyzer/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
google-genai==0.3.0
yt-dlp==2024.10.7
